Move quick start dynamic card related classes to dynamiccards package
